Reorganize types_builtin.go and add yangtype.go and fix YangRange Equal()
- Put `YangType` and related names in yangtype.go.
- Add `Equal` function for `YRange` and fix `YangRange`'s `Equal` function.
